feat(core, persistence, indexer): pool_current_state read model + projection wiring
Introduce a CQRS-style projection of the latest known state of each
pool, maintained event-driven by the indexer. This unlocks O(1)
per-pool lookups for the upcoming pool detail page on the dashboard,
without scanning the swap_events / liquidity_events hypertables.

Type alignment:
  * reserve_a / reserve_b   u64   ↔ BIGINT          (matches event tables)
  * last_sqrt_price         u128  ↔ NUMERIC(39, 0)  (Q64.64)
  * liquidity               u128  ↔ NUMERIC(39, 0)  (concentrated-liquidity L)

* migrations/003_pool_current_state.{up,down}.sql
  - Table pool_current_state with one row per pool, FK to pools
  - CHECK constraint on last_event_kind ('swap' | 'liquidity_add' |
    'liquidity_remove')
  - Indexes on (protocol) and (last_event_at DESC)
  - Idempotent backfill from existing swap_events / liquidity_events,
    picking the most recent event between the two streams per pool
  - Reversible: down migration drops indexes + table

* yog-core (domain)
  - PoolCurrentState (read model)
  - PoolCurrentStateUpsert (write payload), with from_swap /
    from_liquidity constructors. from_liquidity takes the domain
    LiquidityEventKind directly — no bare bool flag
  - LastEventKind enum + wire roundtrip
  - From<LiquidityEventKind> for LastEventKind centralizes the
    add/remove mapping in one place (lives next to LastEventKind
    so LiquidityEventKind stays unaware of the projection)
  - PoolCurrentStateRepository trait (async_trait); contract documents
    stale-write safety and sqrt_price / liquidity preservation semantics

* yog-persistence
  - PgPoolCurrentStateRepository
  - Stale-write guard in SQL via WHERE on the ON CONFLICT DO UPDATE
    branch — out-of-order events are silently ignored
  - last_sqrt_price / last_swap_at preserved on liquidity events,
    liquidity / last_liquidity_at preserved on swap events
  - u64 / u128 conversions go through repository_utils helpers for
    consistency with the existing event repositories

* yog-indexer
  - PoolCurrentStateRepository wired into IndexerService::new
  - Projection upsert runs AFTER each swap_events / liquidity_events
    insert succeeds — current_state only reflects what actually landed
  - Best-effort, skip-and-log: a failure on the projection never
    aborts per-event or per-transaction processing
  - Metrics: persist duration recorded under labels
    `pool_current_state_applied` and `pool_current_state_stale` so
    Prometheus can surface stale-write rates as a health signal

Endpoints in the api crate that consume the projection
(`GET /api/pools/{address}`, `/latest-state`, etc.) come in a
follow-up commit.

+++ crates/persistence/migrations/002_pool_current_state.sql
@@ -1,4 +1,4 @@
--- 003_pool_current_state.sql
+-- 002_pool_current_state.sql
 --
 -- Projection table holding the latest known state of each observed pool.
 --
@@ -21,6 +21,12 @@ BEGIN;
 -- -------------------------------------------------------------------------
 -- Table
 -- -------------------------------------------------------------------------
+--
+-- Column types intentionally mirror their counterparts in swap_events and
+-- liquidity_events:
+--   * reserve_a / reserve_b      BIGINT          (u64 in the domain — SPL amounts)
+--   * last_sqrt_price            NUMERIC(39, 0)  (u128 Q64.64)
+--   * liquidity                  NUMERIC(39, 0)  (u128 concentrated-liquidity L)
 
 CREATE TABLE IF NOT EXISTS pool_current_state (
     pool_address       TEXT PRIMARY KEY
@@ -33,8 +39,8 @@ CREATE TABLE IF NOT EXISTS pool_current_state (
     last_signature     TEXT NOT NULL,
 
     -- Canonical reserves (token_a, token_b ordering as established in pools)
-    reserve_a          NUMERIC(39, 0) NOT NULL,
-    reserve_b          NUMERIC(39, 0) NOT NULL,
+    reserve_a          BIGINT NOT NULL,
+    reserve_b          BIGINT NOT NULL,
 
     -- Price proxy: sqrt_price is updated by swap events only
     last_sqrt_price    NUMERIC(39, 0),
